feat(prompt trimming): add context token size management for prompts
This PR attempts to restructure prompting a bit to allow for more concise prompting that respects known token-size limits of the LLM. As part of this work, various prompt issues are addressed through attempts at various rephrasings and examples.

This PR also turns up the temperature of the story generator, in an attempt to produce more variety in story outcomes, etc. A random number of paragraphs is introduced in an attempt to vary the story pace a tiny bit (unclear if this is ultimately desirable).

The TalkToSomeoneTool is removed from consideration for now.

Small "optimizations" are made (ex: fewer system message update calls, no new files for likelihood blocks, etc.).