QOL: Enable IDE hinting for settings schema reference (#64)
This really underscores that Python has type hinting: settings schema previously relied on accessing the fields of the ServerSettings object by a dict-style lookup, which, being plain-text, doesn't come along with text-completion in IDEs. This introduces the concept of a classmethod schema_instance which on its face is a ServerSettings object, but fields are instantiated with their schemas (i.e., a dict) rather than whatever their default value is.